shift FollowBtn into reusable component
